Do /boot file update after package install
When dpkg tries to update a file in /boot, it first tries to use a
hard link to backup the old version. But this fails on a fat32
partition, because it doesn't support hard links. The initial install
will work, but updates will fail.

The fix is to package things on the linux partition instead, doesn't
matter where, but then use an after-install script to copy them over
to the right spot on /boot after the package installs.