Rename atheros firmware file
The driver will load /lib/firmware/htc_9271.fw but may be looking
in a subdirectory for the one with the version number attached. Either
way we can name it like this, it works. It's not like that firmware
ever gets updated anyway, it's been one patched version for years.

We already block the firware-atheros package, so it won't be updated
and overwritten by anything else.