Stop uploading kernels to the 2.0 repo
The older code depends on several kernel patches that we are going
to drop, for example the patch that renames the Broadcom wifi cards
so that the scripts can tell which interface is for the hotspot.

We don't do that anymore, wifi interfaces are all usable for any
purpose and don't need to be renamed.